requirements.txt
Django==1.10
MySQL-python==1.2.5
future==0.15.2
django-extensions==1.7.2
django-import-export==0.4.5
#django-debug-toolbar==
-e git+https://github.com/django-debug-toolbar/django-debug-toolbar.git#egg=django-debug-toolbar